refactor(frontend): split render view models
