+++ apps/frontend/src/lib/api-client/roomTimeline.ts
@@ -3,7 +3,8 @@ import { notifyUserSummaries } from './hooks.js';
 import { authHeaders, createChattoClient, handleAuthError } from './connect.js';
 import type { RawEvent, EventConnectionPage, UserSummaryForCache } from './events.js';
 import { RoomEventKind } from './eventKinds.js';
-import { type RoomEventView, type SocialPostPreviewView } from './renderTypes.js';
+import type { RoomEventView } from './renderTypes.js';
+import type { SocialPostPreviewView } from '$lib/render/linkPreviews';
 import { MessageService } from '@chatto/api-types/api/v1/messages_connect';
 import { RoomService } from '@chatto/api-types/api/v1/rooms_connect';
 import { ThreadService } from '@chatto/api-types/api/v1/threads_connect';
