+++ apps/frontend/src/lib/render/reactions.ts
@@ -1,0 +1,6 @@
+export type ReactionSummaryView = {
+  emoji: string;
+  count: number;
+  hasReacted: boolean;
+  users: Array<{ id: string; displayName: string }>;
+};
